BaseTools: Fix PcdNvStoreDefaultValueBuffer Value.
V2:
update the commit log. This issue is related to the commit
e6eae3b4c7b9b756263ecec79694de5f1e85b73a
and
0b6c5954e1d9a17e01eee7d5ef840a5b4790e2e8

V1:
https://bugzilla.tianocore.org/show_bug.cgi?id=1385
PcdNvStoreDefaultValueBuffer value is update to Vpd Info File,
but it is not update into a internal cache. This patch is
going to fix this incorrect value in that internal cache.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Bob Feng <bob.c.feng@intel.com>
Cc: Liming Gao <liming.gao@intel.com>